requirements.txt
flask
langchain
anthropic
sentence-transformers
faiss-cpu
python-dotenv
openai